pyproject.toml
[project]
name = "web_vuln_triage"
version = "0.1.0"
description = "Web Vulnerability Triage RL Environment - classify, detect and prioritize vulnerabilities"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.1.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
]

[project.scripts]
server = "web_vuln_triage.server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
]